Обновил репозиторий

+++ src/views/model/list.php
@@ -13,7 +13,7 @@ use janisto\ycm\widgets\Alert;
 $module = Yii::$app->controller->module;
 
 $this->title = $module->getAdminName($model);
-$this->params['breadcrumbs'][] = ['label' => Yii::t('ycm', 'Content'), 'url' => ['index']];
+//$this->params['breadcrumbs'][] = ['label' => Yii::t('ycm', 'Content'), 'url' => ['index']];
 $this->params['breadcrumbs'][] = $this->title;
 
 ?>
@@ -27,7 +27,12 @@ $this->params['breadcrumbs'][] = $this->title;
     <p>
         <?php
         if ($module->getHideCreate($model) === false) {
-            echo Html::a(Yii::t('ycm', 'Create {name}', ['name' => $module->getSingularName($name)]), ['create', 'name' => $name], ['class' => 'btn btn-success']);
+			if(method_exists($model, 'getCreateParams')) {
+				echo Html::a(Yii::t('ycm', 'Create {name}', ['name' => $module->getSingularName($name)]), array_merge(['create', 'name' => $name], $model->createParams), ['class' => 'btn btn-success']);
+			}
+            else {
+				echo Html::a(Yii::t('ycm', 'Create {name}', ['name' => $module->getSingularName($name)]), ['create', 'name' => $name], ['class' => 'btn btn-success']);
+			}
         }
         ?>
     </p>
